feat: start fixing bug
