+++ engine/language_client_typescript/errors.js
@@ -1,7 +1,9 @@
 "use strict";
 // NOTE: Don't take a dependency on ./native here, it will break the browser code
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.toBamlError = exports.isBamlError = exports.BamlClientHttpError = exports.BamlValidationError = exports.BamlClientFinishReasonError = void 0;
+exports.BamlClientHttpError = exports.BamlValidationError = exports.BamlClientFinishReasonError = void 0;
+exports.isBamlError = isBamlError;
+exports.toBamlError = toBamlError;
 class BamlClientFinishReasonError extends Error {
     prompt;
     raw_output;
@@ -154,7 +156,6 @@ function isBamlError(error) {
         error instanceof BamlValidationError ||
         error instanceof BamlClientFinishReasonError);
 }
-exports.isBamlError = isBamlError;
 function toBamlError(error) {
     try {
         if (isBamlError(error)) {
@@ -166,5 +167,4 @@ function toBamlError(error) {
         return error;
     }
 }
-exports.toBamlError = toBamlError;
 // No need for a separate throwBamlValidationError function in TypeScript
