+++ integ-tests/react/package.json
@@ -30,9 +30,11 @@
     "@radix-ui/react-toggle": "^1.1.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "jotai": "2.12.1",
     "lucide-react": "^0.471.1",
     "next": "^15.1.4",
     "next-themes": "^0.4.4",
+    "nuqs": "2.4.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "sonner": "^1.7.1",
